rename auth actions/reducer to user in redux
