+++ webclient/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component, lazy, Suspense } from "react";
 import { connect } from "react-redux";
-import * as actions from "./store/actions/auth";
+import * as actions from "./store/actions/user";
 import { Switch, Redirect, Route, BrowserRouter } from "react-router-dom";
 import Loader from "./components/UI/Loader/Loader";
 import "./App.scss";
@@ -48,7 +48,7 @@ class App extends Component {
 }
 
 const mapStateToProps = state => ({
-  token: state.auth.token
+  token: state.user.token
 });
 
 const mapDispatchToProps = dispatch => ({
