+++ webclient/src/containers/Main/Main.js
@@ -41,7 +41,7 @@ class Main extends Component {
 }
 
 const mapStateToProps = state => ({
-  extensions: state.auth.user.extensions
+  extensions: state.user.user.extensions
 });
 
 const mapDispatchToProps = dispatch => ({
